(bug): fix drawer animations, fix close all bug, hookup stacking to dialogs

+++ src/components/dialog/dialog.service.ts
@@ -21,6 +21,7 @@ export class DialogService extends InjectionRegistery {
     }
   };
 
+  zIndex: number = 995;
   private closeSubscription;
 
   constructor(
@@ -31,28 +32,59 @@ export class DialogService extends InjectionRegistery {
 
   create(bindings): any {
     const component = super.create(bindings);
-    const { instance } = component;
-
-    this.closeSubscription = instance.close.subscribe(() => {
-      this.destroy(component);
-    });
-
-    if(instance.showOverlay) {
-      this.overlayService.show();
-
-      if(instance.closeOnBlur) {
-        let overlayListener = this.overlayService.instance.click.subscribe(() => {
-          this.destroy(component);
-        });
-      }
-    }
-
+    this.createSubscriptions(component);
     return component;
   }
 
-  destroy(instance) {
-    super.destroy(instance);
-    this.overlayService.destroy();
+  destroy(component): void {
+    const hasOverlay = component.instance.showOverlay;
+    super.destroy(component);
+    this.zIndex = this.zIndex - 1;
+
+    if(hasOverlay) {
+      const compsByType = this.getByType();
+      if(!compsByType.length) {
+        this.overlayService.destroy();
+      } else {
+        this.overlayService.instance.zIndex = this.zIndex - 1; 
+      }
+    }
+  }
+
+  createSubscriptions(component): any {
+    let closeSub;
+    let overlaySub;
+
+    const kill = () => {
+      const compsByType = this.getByType();
+      const lastComp = compsByType[compsByType.length - 1];
+
+      if(lastComp === component) {
+        closeSub.unsubscribe();
+        if(overlaySub) overlaySub.unsubscribe();
+        this.destroy(component);
+      }
+    };
+
+    closeSub = component.instance.close.subscribe(kill);
+
+    if(component.instance.showOverlay) {
+      const overlay = this.overlayService.show();
+      if(component.instance.closeOnBlur) {
+        overlaySub = this.overlayService.instance.click.subscribe(kill);
+      }
+    }
+  }
+
+  assignDefaults(bindings): any {
+    bindings = super.assignDefaults(bindings);
+
+    if(!bindings.zIndex) {
+      this.zIndex = this.zIndex + 1;
+      bindings.inputs.zIndex = this.zIndex;
+    }
+    
+    return bindings;
   }
 
 }
